Restructure generate command
